+++ libs/core/langchain_core/sql_constructor/schema.py
@@ -1,0 +1,15 @@
+from langchain_core.pydantic_v1 import BaseModel
+
+
+class AttributeInfo(BaseModel):
+    """Information about a data source attribute."""
+
+    name: str
+    description: str
+    type: str
+
+    class Config:
+        """Configuration for this pydantic object."""
+
+        arbitrary_types_allowed = True
+        frozen = True
