+++ libs/langchain/langchain/retrievers/self_query/pgvector.py
@@ -1,52 +1,3 @@
-from typing import Dict, Tuple, Union
+from langchain_community.retrievers.self_query.pgvector import PGVectorTranslator
 
-from langchain.chains.query_constructor.ir import (
-    Comparator,
-    Comparison,
-    Operation,
-    Operator,
-    StructuredQuery,
-    Visitor,
-)
-
-
-class PGVectorTranslator(Visitor):
-    """Translate `PGVector` internal query language elements to valid filters."""
-
-    allowed_operators = [Operator.AND, Operator.OR]
-    """Subset of allowed logical operators."""
-    allowed_comparators = [
-        Comparator.EQ,
-        Comparator.NE,
-        Comparator.GT,
-        Comparator.LT,
-        Comparator.IN,
-        Comparator.NIN,
-        Comparator.CONTAIN,
-        Comparator.LIKE,
-    ]
-    """Subset of allowed logical comparators."""
-
-    def _format_func(self, func: Union[Operator, Comparator]) -> str:
-        self._validate_func(func)
-        return f"{func.value}"
-
-    def visit_operation(self, operation: Operation) -> Dict:
-        args = [arg.accept(self) for arg in operation.arguments]
-        return {self._format_func(operation.operator): args}
-
-    def visit_comparison(self, comparison: Comparison) -> Dict:
-        return {
-            comparison.attribute: {
-                self._format_func(comparison.comparator): comparison.value
-            }
-        }
-
-    def visit_structured_query(
-        self, structured_query: StructuredQuery
-    ) -> Tuple[str, dict]:
-        if structured_query.filter is None:
-            kwargs = {}
-        else:
-            kwargs = {"filter": structured_query.filter.accept(self)}
-        return structured_query.query, kwargs
+__all__ = ["PGVectorTranslator"]
